e_paper_calendar/calendar_server_side/requirements.txt
Pillow
requests
numpy
webdavclient3
